+++ src/platform_impl/web/event_loop/proxy.rs
@@ -3,18 +3,18 @@ use std::sync::mpsc::Sender;
 use super::runner;
 use crate::event::Event;
 use crate::event_loop::EventLoopClosed;
-use crate::platform_impl::platform::r#async::Channel;
+use crate::platform_impl::platform::r#async::Waker;
 
 pub struct EventLoopProxy<T: 'static> {
     // used to wake the event loop handler, not to actually pass data
-    runner: Channel<runner::Shared, ()>,
+    runner: Waker<runner::Shared, ()>,
     sender: Sender<T>,
 }
 
 impl<T: 'static> EventLoopProxy<T> {
     pub fn new(runner: runner::Shared, sender: Sender<T>) -> Self {
         Self {
-            runner: Channel::new(runner, |runner, event| {
+            runner: Waker::new(runner, |runner, event| {
                 runner.send_event(Event::UserEvent(event))
             })
             .unwrap(),
